+++ src/yuni/core/system/environment.h
@@ -22,15 +22,26 @@ namespace Environment
 
 
 	/*!
-	** \brief Get a value from the current environment
+	** \brief Get a value from the current environment (Yuni::String version)
 	**
 	** \param name Name of the variable
 	** \param[out] out Variable of type 'string'/'container' where the value will be appened
 	** \param emptyBefore True to empty the parameter `out` before
 	** \return True if the operation succeeded, false otherwise
 	*/
-	template<class StringT>
-	YUNI_DECL bool Read(const AnyString& name, StringT& out, bool emptyBefore = true);
+	YUNI_DECL bool Read(const AnyString& name, String& out, bool emptyBefore = true);
+
+	/*!
+	** \brief Get a value from the current environment (Yuni::Clob version)
+	**
+	** \param name Name of the variable
+	** \param[out] out Variable of type 'string'/'container' where the value will be appened
+	** \param emptyBefore True to empty the parameter `out` before
+	** \return True if the operation succeeded, false otherwise
+	*/
+	YUNI_DECL bool Read(const AnyString& name, Clob& out, bool emptyBefore = true);
+
+
 
 
 
@@ -39,7 +50,5 @@ namespace Environment
 } // namespace Environment
 } // namespace System
 } // namespace Yuni
-
-# include "environment.hxx"
 
 #endif // __YUNI_CORE_SYSTEM_ENVIRONMENT_H__